requirements.txt
google-genai>=1.0.0
streamlit>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.0
